Use cobra. Add 'up' subcommand
## Summary

- Use cobra to structure commands
- Add the `up` subcommand to sync (`jj git fetch` + `jj rebase -d main|master` + drop empty commits)
