+++ main.go
@@ -1,227 +1,13 @@
 package main
 
 import (
-	"fmt"
 	"os"
-	"os/exec"
-	"path/filepath"
-	"strconv"
-	"strings"
 
-	"github.com/knadh/koanf/parsers/toml"
-	"github.com/knadh/koanf/providers/confmap"
-	"github.com/knadh/koanf/providers/file"
-	"github.com/knadh/koanf/v2"
-	"golang.org/x/exp/slog"
+	"github.com/lazywei/fj/cmd"
 )
 
-var k = *koanf.New(".")
-
-func InitConfig() {
-	// Load default values using the confmap provider.
-	// We provide a flat map with the "." delimiter.
-	// A nested map can be loaded by setting the delimiter to an empty string "".
-	k.Load(confmap.Provider(map[string]interface{}{
-		"mainBranch":   "main",
-		"branchPrefix": "username/pr-",
-	}, "."), nil)
-
-	projectRoot, err := Run(exec.Command("git", "rev-parse", "--show-toplevel"))
-	if err != nil {
-		slog.Error("Not in a git repo", "err", err)
-		os.Exit(1)
-	}
-	configPath := filepath.Join(projectRoot, ".fj.toml")
-
-	// Load TOML config on top of the default values.
-	if err := k.Load(file.Provider(configPath), toml.Parser()); err != nil {
-		b, err := k.Marshal(toml.Parser())
-		if err != nil {
-			slog.Error("Failed to marshal default config", "err", err)
-			os.Exit(1)
-		}
-		if err := os.WriteFile(configPath, b, 0o644); err != nil {
-			slog.Error("Failed to write default config", "configPath", configPath, "err", err)
-			os.Exit(1)
-		}
-		slog.Info("Initialized default config, please re-run the command", "configPath", configPath)
-		os.Exit(0)
-	}
-}
-
 func main() {
-	InitConfig()
-
-	changeIDs, err := getStackChangeIDs()
-	if err != nil {
-		slog.Error("Failed to get Change IDs for the stack", "err", err)
+	if err := cmd.Execute(); err != nil {
 		os.Exit(1)
 	}
-
-	prStack := []int{}
-	descriptions := []string{}
-
-	lastBranch := ""
-	for _, changeID := range changeIDs {
-		desc, err := getDescription(changeID)
-		if err != nil {
-			slog.Error("Failed to get description", "change", changeID, "err", err)
-			os.Exit(1)
-		}
-
-		branch, err := getBranch(changeID)
-		if err != nil {
-			slog.Error("Failed to get branch", "change", changeID, "err", err)
-			os.Exit(1)
-		}
-
-		if branch == "" {
-			nextPRNum, err := getNextAvailablePRNumber()
-			if err != nil {
-				slog.Error("Failed to getNextAvailablePRNumber", "err", err)
-				os.Exit(1)
-			}
-
-			branch, err = createBranch(changeID, nextPRNum)
-			if err != nil {
-				slog.Error("Failed to create branch", "err", err)
-				os.Exit(1)
-			}
-		}
-
-		if err := exec.Command("jj", "git", "push", "-r", changeID).Run(); err != nil {
-			slog.Error("Failed to push branch to remote", "err", err)
-			os.Exit(1)
-		}
-		slog.Debug("Branch pushed to remote")
-
-		prNum, err := getPRNumber(branch)
-		if err != nil {
-			slog.Error("Failed to check PR exisitence", "err", err)
-			os.Exit(1)
-		}
-
-		if prNum == -1 {
-			slog.Debug("No PR created for branch yet, creating", "branch", branch)
-
-			var baseBranch string
-			if lastBranch == "" {
-				baseBranch = k.String("mainBranch")
-			} else {
-				baseBranch = lastBranch
-			}
-
-			out, err := Run(exec.Command("gh", "pr", "create", "-H", branch, "-B", baseBranch, "--fill-first"))
-			if err != nil {
-				slog.Error("Failed to create PR", "output", out, "err", err)
-				os.Exit(1)
-			}
-			slog.Debug("PR created", "output", out)
-			prNum, _ = getPRNumber(branch)
-		}
-		prStack = append(prStack, prNum)
-		descriptions = append(descriptions, desc)
-
-		lastBranch = branch
-	}
-
-	for i, prNum := range prStack {
-		desc := descriptions[i]
-
-		prInfo := "\n---"
-		for j := len(prStack) - 1; j >= 0; j-- {
-			if i == j {
-				prInfo += fmt.Sprintf("\n* **->** #%d", prStack[j])
-			} else {
-				prInfo += fmt.Sprintf("\n* #%d", prStack[j])
-			}
-		}
-
-		if out, err := Run(exec.Command("gh", "pr", "edit", fmt.Sprint(prNum), "-b", desc+"\n"+prInfo)); err == nil {
-			fmt.Println("Successfully updated PR:", out)
-		}
-	}
-}
-
-func Run(cmd *exec.Cmd) (string, error) {
-	out, err := cmd.CombinedOutput()
-	ret := strings.TrimRight(string(out), "\n")
-	return ret, err
-}
-
-func getStackChangeIDs() ([]string, error) {
-	out, err := Run(exec.Command(
-		"jj", "log", "--no-graph", "--reversed",
-		"-r", fmt.Sprintf("%s..@-", k.String("mainBranch")),
-		"-T", `change_id ++ "\n"`,
-	))
-	if err != nil {
-		return []string{}, fmt.Errorf("output: %s, err: %v", out, err)
-	}
-
-	changeIDs := strings.Split(out, "\n")
-	return changeIDs, nil
-}
-
-func getDescription(changeID string) (string, error) {
-	return Run(exec.Command(
-		"jj", "log", "--no-graph",
-		"-T", `description`,
-		"-r", changeID,
-	))
-}
-
-func getBranch(changeID string) (string, error) {
-	out, err := Run(exec.Command(
-		"jj", "branch", "list",
-		"-r", changeID,
-	))
-	if err != nil {
-		return out, err
-	}
-	return strings.Split(out, ":")[0], nil
-}
-
-func getNextAvailablePRNumber() (int, error) {
-	cmd := "gh pr list -L 1 --state all --json number | jq '.[0].number'"
-	out, err := Run(exec.Command("bash", "-c", cmd))
-	if err != nil {
-		return -1, fmt.Errorf("output: %s, err: %v", out, err)
-	}
-	if out == "null" {
-		// No PRs yet
-		return 1, nil
-	}
-	curPRNum, err := strconv.Atoi(out)
-	if err != nil {
-		return -1, err
-	}
-
-	return curPRNum + 1, nil
-}
-
-func createBranch(changeID string, nextPRNum int) (string, error) {
-	branchName := fmt.Sprintf("cwc/pr-%d", nextPRNum)
-
-	out, err := Run(exec.Command("jj", "branch", "create", "-r", changeID, branchName))
-	if err != nil {
-		return "", fmt.Errorf("output: %s, err: %v", out, err)
-	}
-
-	return branchName, nil
-}
-
-func getPRNumber(branch string) (int, error) {
-	cmd := fmt.Sprintf("gh pr list -L 1 --state all --json number --head %s | jq '.[0].number'", branch)
-	out, err := Run(exec.Command("bash", "-c", cmd))
-	if err != nil {
-		return -1, err
-	}
-
-	number, err := strconv.Atoi(out)
-	if err != nil {
-		return -1, nil
-	}
-
-	return number, nil
 }
